# api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: Transaction validation service with rules and limits
  termsOfService: http://swagger.io/terms/
  title: Tracer API
  version: 0.1.0
servers:
- url: http://localhost:8080/
- url: https://localhost:8080/
paths:
...
  /v1/limits:
    get:
      description: Lists limits with cursor-based pagination and optional filters.
      operationId: listLimits
      parameters:
      - description: "Max items per page (1-100, default: 10)"
        in: query
        name: limit
        schema:
          maximum: 100
          minimum: 1
          type: integer
      - description: Pagination cursor (empty for first page)
        in: query
        name: cursor
        schema:
          type: string
      - description: Filter by status
        in: query
        name: status
        schema:
          enum:
          - DRAFT
          - ACTIVE
          - INACTIVE
          type: string
      - description: Filter by limit type
        in: query
        name: limitType
        schema:
          enum:
          - DAILY
          - MONTHLY
          - PER_TRANSACTION
          type: string
      - description: Sort field
        in: query
        name: sortBy
        schema:
          enum:
          - createdAt
          - updatedAt
          - name
          - maxAmount
          type: string
      - description: Sort direction
        in: query
        name: sortOrder
        schema:
          enum:
          - ASC
          - DESC
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapters_http_in.ListLimitsResponse'
          description: Limits listed successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid parameters
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: List spending limits
      tags:
      - limits
    post:
      description: Creates a limit with scopes array. Limits are created in DRAFT
        status.
      operationId: createLimit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/internal_adapters_http_in.CreateLimitInput'
        description: Limit details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tracer_pkg_model.Limit'
          description: Limit created successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid input
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Limit name already exists
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: Create a new spending limit
      tags:
      - limits
      x-codegen-request-body-name: limit
  /v1/limits/{id}:
    delete:
      description: Soft-deletes a limit (transitions to DELETED status). Only DRAFT
        and INACTIVE limits can be deleted. ACTIVE limits must be deactivated first.
      operationId: deleteLimit
      parameters:
      - description: Limit ID (UUID)
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "204":
          content: {}
          description: Limit deleted successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid limit ID or transition
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Limit not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: Delete a spending limit
      tags:
      - limits
    get:
      description: Retrieves a limit by its unique identifier.
      operationId: getLimit
      parameters:
      - description: Limit ID (UUID)
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tracer_pkg_model.Limit'
          description: Limit retrieved successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid limit ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Limit not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: Get a spending limit by ID
      tags:
      - limits
    patch:
      description: "Partially updates a limit. Only provided fields are updated, omitted\
        \ fields remain unchanged."
      operationId: updateLimit
      parameters:
      - description: Limit ID (UUID)
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/internal_adapters_http_in.UpdateLimitInput'
        description: Fields to update
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tracer_pkg_model.Limit'
          description: Limit updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid input
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Limit not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: Partially update an existing spending limit
      tags:
      - limits
      x-codegen-request-body-name: limit
  /v1/limits/{id}/activate:
    post:
      description: Activates an inactive limit (INACTIVE → ACTIVE).
      operationId: activateLimit
      parameters:
      - description: Limit ID (UUID)
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tracer_pkg_model.Limit'
          description: Limit activated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid limit ID or transition
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Limit not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: Activate a spending limit
      tags:
      - limits
...
components:
  schemas:
...
    internal_adapters_http_in.ListLimitsResponse:
      example:
        nextCursor: nextCursor
        hasMore: true
        limits:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          deletedAt: 2000-01-23T04:56:07.000+00:00
          resetAt: 2000-01-23T04:56:07.000+00:00
          name: name
          description: description
          limitId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          currency: currency
          scopes:
          - transactionType: CARD
            accountId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            portfolioId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            merchantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            segmentId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            subType: subType
          - transactionType: CARD
            accountId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            portfolioId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            merchantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            segmentId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            subType: subType
          maxAmount: 100000
          limitType: DAILY
          status: DRAFT
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          deletedAt: 2000-01-23T04:56:07.000+00:00
          resetAt: 2000-01-23T04:56:07.000+00:00
          name: name
          description: description
          limitId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          currency: currency
          scopes:
          - transactionType: CARD
            accountId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            portfolioId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            merchantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            segmentId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            subType: subType
          - transactionType: CARD
            accountId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            portfolioId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            merchantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            segmentId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            subType: subType
          maxAmount: 100000
          limitType: DAILY
          status: DRAFT
          updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        hasMore:
          type: boolean
        limits:
          items:
            $ref: '#/components/schemas/tracer_pkg_model.Limit'
          type: array
        nextCursor:
          type: string
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: API Key for authentication
      in: header
      name: X-API-Key
      type: apiKey
...